Edit README.md and improve code quality
