+++ src/main/java/yoyo/core/Launcher.java
@@ -1,4 +1,4 @@
-package yoyo;
+package yoyo.core;
 
 import javafx.application.Application;
