+++ src/main/java/yoyo/core/Main.java
@@ -1,4 +1,4 @@
-package yoyo;
+package yoyo.core;
 
 import java.io.IOException;
 
@@ -7,6 +7,7 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
+import yoyo.ui.MainWindow;
 
 /**
  * A GUI for Yoyo using FXML.
