+++ src/main/java/yoyo/task/Event.java
@@ -1,12 +1,10 @@
-package yoyo;
+package yoyo.task;
 
-import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
-import java.util.Arrays;
-import java.util.Optional;
+
+import yoyo.util.Constants;
+import yoyo.util.DateTimeUtil;
 
 /**
  * Represents an event task with a start and end time.
@@ -28,8 +26,8 @@ public class Event extends Task {
         super(TaskType.EVENT, description);
         assert fromRaw != null && !fromRaw.trim().isEmpty() : "Event 'from' parameter cannot be null or empty";
         assert toRaw != null && !toRaw.trim().isEmpty() : "Event 'to' parameter cannot be null or empty";
-        this.from = parseFlexibleDateTime(fromRaw);
-        this.to = parseFlexibleDateTime(toRaw);
+        this.from = DateTimeUtil.parseFlexibleDateTime(fromRaw);
+        this.to = DateTimeUtil.parseFlexibleDateTime(toRaw);
         if (this.to.isBefore(this.from)) {
             throw new IllegalArgumentException("Event end time must be after start time.");
         }
@@ -78,59 +76,5 @@ public class Event extends Task {
                 description,
                 from.format(STORAGE_FORMAT),
                 to.format(STORAGE_FORMAT));
-    }
-
-    /**
-     * Parses a flexible date/time string into a LocalDateTime. Supports various
-     * formats like yyyy-MM-dd, d/M/yyyy, with optional time.
-     *
-     * @param raw the raw date/time string
-     * @return the parsed LocalDateTime
-     * @throws IllegalArgumentException if parsing fails
-     */
-    private static LocalDateTime parseFlexibleDateTime(String raw) {
-        String s = raw.trim();
-
-        // Try datetime patterns first using streams
-        Optional<LocalDateTime> dateTimeResult = Arrays.stream(new DateTimeFormatter[]{
-            DateTimeFormatter.ofPattern(Constants.DATETIME_FORMAT_STORAGE),
-            DateTimeFormatter.ofPattern(Constants.DATE_FORMAT_SHORT + " HHmm")
-        })
-        .map(formatter -> {
-            try {
-                return LocalDateTime.parse(s, formatter);
-            } catch (DateTimeParseException e) {
-                return null;
-            }
-        })
-        .filter(result -> result != null)
-        .findFirst();
-
-        if (dateTimeResult.isPresent()) {
-            return dateTimeResult.get();
-        }
-
-        // Then try date-only (assume 00:00)
-        Optional<LocalDateTime> dateResult = Arrays.stream(new DateTimeFormatter[]{
-            DateTimeFormatter.ISO_LOCAL_DATE, // yyyy-MM-dd
-            DateTimeFormatter.ofPattern(Constants.DATE_FORMAT_SHORT) // 2/12/2019
-        })
-        .map(formatter -> {
-            try {
-                LocalDate d = LocalDate.parse(s, formatter);
-                return LocalDateTime.of(d, LocalTime.MIDNIGHT);
-            } catch (DateTimeParseException e) {
-                return null;
-            }
-        })
-        .filter(result -> result != null)
-        .findFirst();
-
-        if (dateResult.isPresent()) {
-            return dateResult.get();
-        }
-
-        throw new IllegalArgumentException(
-                String.format(Constants.ERR_UNRECOGNIZED_DATE, raw));
     }
 }
